pyproject.toml
[tool.poetry]
name = "sm2t_traffic_model"
version = "0.1.0"
description = ""
authors = ["Christina Ludwig <christina.ludwig@uni-heidelberg.de>"]

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^0.12"
requests = "^2.27.1"
tqdm = "^4.62.3"
urllib3 = "^1.26.8"
scikit-learn = "^1.0.2"
seaborn = "^0.12"
joblib = "^1.1.0"
SQLAlchemy = "^1.4.31"
GeoAlchemy2 = "^0.10.2"
psycopg2 = "^2.9.3"
python-dotenv = "^1.0"
SQLAlchemy-Utils = "^0.38.2"
ipykernel = "^6.9.1"
xgboost = "^1.7.4"
utm = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
pre-commit = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
